requirements.txt
discord.py
chat_exporter
python-dotenv